% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{blosc_compress}
\alias{blosc_compress}
\title{Compress with BLOSC}
\usage{
blosc_compress(
  x,
  compressor = "lz4",
  level = 7,
  shuffle = "shuffle",
  typesize = 4,
  ...
)
}
\arguments{
\item{x}{TODO}

\item{compressor}{todo}

\item{level}{TODO}

\item{shuffle}{TODO}

\item{typesize}{TODO}

\item{...}{Ignored}
}
\description{
TODO
}
\examples{
my_dat      <- as.raw(sample.int(255, 10*1024, replace = TRUE))
con         <- rawConnection(my_dat)
my_dat_out1 <- blosc_compress(my_dat)
#my_dat_out2 <- blosc_compress(con) #TODO

}
