% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress.R
\name{blosc_info}
\alias{blosc_info}
\title{Information about compressed data}
\usage{
blosc_info(x, ...)
}
\arguments{
\item{x}{Raw data compressed with \code{blosc_compress()}.}

\item{...}{Ignored}
}
\value{
Returns a named list with information about blosc compressed
data \code{x}.
}
\description{
Obtain information about raw data compressed with blosc.
}
\examples{
data_compressed <-
  blosc_compress(volcano, typesize = 2, dtype = "<i2", compressor = "lz4",
                 shuffle = "bitshuffle")

blosc_info(data_compressed)
}
