% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosc-package.R
\docType{package}
\name{blosc-package}
\alias{blosc}
\alias{blosc-package}
\title{blosc: Compress and Decompress Data Using the BLOSC Library}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Blosc is a blocking, shuffling and lossless compression algorithm. It is often used to compress n-dimensional structured array data. It is applied (amongst others) in the geo-spatial zarr file format. This package can be used to decompress such data or to compress data.
}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

}
\keyword{internal}
