% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blosc-package.R
\docType{package}
\name{blosc-package}
\alias{blosc}
\alias{blosc-package}
\title{blosc: Compress and Decompress Data Using the BLOSC Library}
\description{
TODO.
}
\author{
\strong{Maintainer}: Pepijn de Vries \email{pepijn.devries@outlook.com} (\href{https://orcid.org/0000-0002-7961-6646}{ORCID})

}
\keyword{internal}
