% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dtype.R
\name{r_to_dtype}
\alias{r_to_dtype}
\alias{dtype_to_r}
\title{Convert from or to ZARR dtypes}
\usage{
r_to_dtype(x, dtype, na_value = NA, ...)

dtype_to_r(x, dtype, na_value = NA, ...)
}
\arguments{
\item{x}{TODO}

\item{dtype}{TODO}

\item{na_value}{TODO}

\item{...}{Ignored}
}
\value{
TODO
}
\description{
TODO description
}
\examples{
r_to_dtype(1:100, "<i4")
r_to_dtype(c(TRUE, FALSE), "|b1")
dtype_to_r(as.raw(0:1), "|b1") #TODO implement
}
\author{
Pepijn de Vries
}
